requirements.txt
fastapi
uvicorn
ultralytics
opencv-python
python-multipart
numpy
pillow
opencv-python-headless
dill